[bugfix] inject vagrant `bin` path to ENV['PATH'] (#48)
* [bugfix] inject vagrant `bin` path to ENV['PATH']

this fixes an issue that Jenkinsfile cannot be executed in slaves.

[packer-template] Running shell script
+ bundle exec rake reallyenglish:clean
/usr/local/lib/ruby/gems/2.2/gems/bundler-1.15.0/lib/bundler/rubygems_integration.rb:432:in
`block in replace_bin_path': can't find executable vagrant for gem vagrant.
vagrant is not currently included in the bundle, perhaps you meant to add it to
your Gemfile? (Gem::Exception)

* [bugfix] remove race condition

using sed and writing to the file being edited has a race condition that
corrupts the file.